fix: share network_transport between coreMQTT and coreHTTP (CA-348)
Extract the ESP-TLS transport into libraries/common/network_transport so
both components link one implementation. This removes duplicate symbol
errors when MQTT and HTTP are used together, keeps the richer MQTT
transport (timeouts/non-blocking I/O), and clears pxTls after disconnect.

Closes #232
Related: #249
