+++ examples/fleet_provisioning/fleet_provisioning_with_csr/CMakeLists.txt
@@ -8,6 +8,7 @@ set(EXTRA_COMPONENT_DIRS "$ENV{IDF_PATH}/examples/common_components/protocol_exa
 						 "${CMAKE_CURRENT_LIST_DIR}/../../../libraries/corePKCS11"
 						 "${CMAKE_CURRENT_LIST_DIR}/../../../libraries/Fleet-Provisioning-for-AWS-IoT-embedded-sdk"
 						 "${CMAKE_CURRENT_LIST_DIR}/../../../libraries/common/posix_compat"
+						 "${CMAKE_CURRENT_LIST_DIR}/../../../libraries/common/network_transport"
 	)
 
 include($ENV{IDF_PATH}/tools/cmake/project.cmake)
