+++ examples/mqtt/tls_mutual_auth/CMakeLists.txt
@@ -6,6 +6,7 @@ set(EXTRA_COMPONENT_DIRS "$ENV{IDF_PATH}/examples/common_components/protocol_exa
 						 "${CMAKE_CURRENT_LIST_DIR}/../../../libraries/backoffAlgorithm"
 						 "${CMAKE_CURRENT_LIST_DIR}/../../../libraries/coreMQTT"
 						 "${CMAKE_CURRENT_LIST_DIR}/../../../libraries/common/posix_compat"
+						 "${CMAKE_CURRENT_LIST_DIR}/../../../libraries/common/network_transport"
 	)
 
 include($ENV{IDF_PATH}/tools/cmake/project.cmake)
