+++ libraries/common/network_transport/CMakeLists.txt
@@ -1,0 +1,8 @@
+idf_component_register(
+    SRCS
+        "network_transport.c"
+    INCLUDE_DIRS
+        "."
+    REQUIRES
+        esp-tls
+)
